backend/requirements.txt
Django==2.2.16
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
